pyproject.toml
[tool.poetry]
name = "backend-final"
version = "0.1.0"
description = ""
authors = ["Mufan <alex1151212@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.88.0"}
motor = "^3.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
